requirements.txt
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.136.0
chromadb==0.5.3
sentence-transformers==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.2.6
langchain_groq==0.1.5
